week2/attention_visualization/requirements.txt
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0